requirements.txt
click==6.7
Jinja2==2.10
peewee==3.0.18
python-dateutil==2.6.1
markdown2==2.3.7
beautifulsoup4==4.7.1
colour-runner==0.1.1
Pygments==2.3.1
python-dotenv==0.10.1
mysqlclient===1.3.14
git+https://github.com/yevrah/simplerr.git
